renamed all components to reflect their imports
