+++ src/components/bio/BioContainer.tsx
