+++ src/components/index.tsx
@@ -1,6 +1,6 @@
-export * from './bio'
-export * from './contact'
-export * from './landing'
-export * from './nav-bar'
-export * from './projects'
-export * from './tech-stack'
+export * from './bio/BioContainer'
+export * from './contact/ContactContainer'
+export * from './landing/LandingContainer'
+export * from './nav-bar/NavBar'
+export * from './projects/ProjectsContainer'
+export * from './tech-stack/TechStackContainer'
